pascal/highscores: Split game.State object
This object did two things before: it was where the state of things
inside the game world lived, like the character and gems. It also
contained state of the program itself, like the logger and window. These
are now split into game.State and states.State.

I think the states.State and phases.Phase object names are kind of dumb,
but I can't think of anything better at the moment.