pascal/highscores: Get scores from DB instead of dummy data
This adds the first dynamic array, which I allocate on an arena. This
crashes the first time scores.State.scores comes into scope because the
runtime tries to add a reference in the reference count system, but my
arena-allocated array is not a managed type. I don't see a way around
this when using Free Pascal's fancy dynamic arrays, so I'll just have to
implement my own dynamic array type.

Fun fact: you can catch an exception like this in GDB to inspect the
stack trace by adding a breakpoint on fpc_raiseexception.