pascal/highscores: Use sqlite3 directly instead of SQLDB
Free Pascal's SQLDB interface is more complicated than I need, and I
didn't understand what it was doing internally. Something I was doing
was leading to database lock errors whenever I tried to read from the
database. The docs should have made it clear how to use the interface
and what I was doing wrong, but the docs are pretty minimal.

As a nice bonus, release builds are an order of magnitude smaller, down
from ~1 MiB to ~100 KiB.