Replace custom log level code with a native SDL hint
The SDL_LOGGING hint was added in SDL 2.30. It does the same thing as my
old LOG_LEVEL env var, but is more flexible.

SDL 2.30 only came out this year, so it's not available everywhere yet.
In older versions of SDL, setting the SDL_LOGGING env var does nothing,
and it's impossible to set the log level in highscores.